[B/ARTICLES] Added option to submit influences while creating an article

+++ dto/articles.go
@@ -34,7 +34,7 @@ type CreateInfluenceParams struct {
 	ArticleID       int32
 	CreatorID       string
 	DurationSeconds int32
-	PermillePerDay  float32
+	PermillePerDay  float64
 	FalloffType     int32
 }
